改善++.

+++ setup.sh
@@ -1,11 +1,16 @@
 #!/usr/bin/env bash
-git clone https://github.com/wujuguang/kyvim.git ~/kyvim/
 
-mv ~/.vimrc ~/.vimrc.bck
-mv ~/.vim ~/.vim.bck
+base_path=$(cd `dirname $0`; pwd)
+git clone https://github.com/wujuguang/kyvim.git ${base_path}/kyvim/
 
-ln -s ~/kyvim/vimrc ~/.vimrc
-ln -s ~/kyvim/vim ~/.vim
+[ -f ~/.vim] && mv ~/.vim ~/.vim.bck
+[ -f ~/.vimrc] && mv ~/.vimrc ~/.vimrc.bck
 
-git clone https://github.com/gmarik/vundle.git ~/kyvim/vim/bundle/vundle/
+ln -s ${base_path}/kyvim/vimrc ~/.vimrc
+ln -s ${base_path}/kyvim/vim ~/.vim
+
+git clone https://github.com/gmarik/vundle.git ${base_path}/kyvim/vim/bundle/vundle/
 vim +PluginInstall +qall
+
+sudo apt-get install -y ctags cscope
+sudo pip install frosted pep8 mccabe
